reorganiza arquivos de build e remove getThemeToken
