+++ index.ts
@@ -1,3 +1,2 @@
 export * from './tokens';
 export * from './token-presets';
-export * from './themes/getThemeToken';
